pyproject.toml
[project]
name = "practica-apis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.10",
]
